Bugs fixed

+++ src/ru/ifmo/ctddev/bandarchuk/expression/Variable.java
@@ -23,7 +23,7 @@ public class Variable implements Operand {
     }
 
     @Override
-    public int evaluate(int x, int y, int z) {
+    public double evaluate(double x, double y, double z) {
         switch (variableName) {
             case "x":
                 return x;
